[FEATURE] Generate server side prepared statements
Add method Papaya\Database\Statement\Prepared::getPreparedSQL()
to generate a server side prepared statement. It returns an
array with the SQL string and the placeholder values.

If possible the original named placeholders in the SQL statement
are replaced by `:?` placeholders (multiple for lists).

Placeholder that are not supported by server side statements
(table name, limit) are replaced with their actual value.