[FEATURE] Allow Content Teasers To Replace Existing Images
Content teasers provide lists of teasers (for example for categories).
The size of the teaser images is should usually be the same for all
teasers in a list. Before the content teasers object added separate
nodes with the thumbnail images. Now you can request it to replace
the original teaser image nodes.